+++ Debaser.Tests/Customization/TestDebaserMapper.cs
@@ -16,8 +16,8 @@ namespace Debaser.Tests.Customization
         {
             _upsertHelper = new UpsertHelper<RowWithJson>("db");
 
-            _upsertHelper.DropSchema();
-            _upsertHelper.CreateSchema();
+            _upsertHelper.DropSchema(dropTable: true, dropProcedure: true, dropType: true);
+            _upsertHelper.CreateSchema(createType: true, createTable: true, createProcedure: true);
         }
 
         [Test]