+++ Debaser.Tests/Query/TestQueries.cs
@@ -13,8 +13,8 @@ namespace Debaser.Tests.Query
         {
             _upsertHelper = new UpsertHelper<RowWithData>("db");
 
-            _upsertHelper.DropSchema();
-            _upsertHelper.CreateSchema();
+            _upsertHelper.DropSchema(dropTable: true, dropProcedure: true, dropType: true);
+            _upsertHelper.CreateSchema(createType: true, createTable: true, createProcedure: true);
         }
 
         [Test]