+++ Debaser.Tests/Readme/ValueType.cs
@@ -15,8 +15,8 @@ namespace Debaser.Tests.Readme
         protected override void SetUp()
         {
             _upsertHelper = new UpsertHelper<CurrencyCrossRates>(ConnectionString);
-            _upsertHelper.DropSchema();
-            _upsertHelper.CreateSchema();
+            _upsertHelper.DropSchema(dropTable: true, dropProcedure: true, dropType: true);
+            _upsertHelper.CreateSchema(createType: true, createTable: true, createProcedure: true);
         }
 
         [Test]